HTTP functions should go to separate file
