+++ tfstate/remote_http.go
@@ -1,0 +1,14 @@
+package tfstate
+
+import (
+	"io"
+	"net/http"
+)
+
+func readHTTP(u string) (io.ReadCloser, error) {
+	resp, err := http.Get(u)
+	if err != nil {
+		return nil, err
+	}
+	return resp.Body, nil
+}
